changed double to decimal as that was giving nan

+++ hardware/stepperCalculator/GCodeData.cs
@@ -4,7 +4,7 @@ namespace stepperCalculator
 {
     public class GCodeData
     {
-        public  Dictionary<PrinterAxis, double> Coordinates = new Dictionary<PrinterAxis, double>
+        public  Dictionary<PrinterAxis, decimal> Coordinates = new Dictionary<PrinterAxis, decimal>
         {
             {PrinterAxis.X, -90000},
             {PrinterAxis.Y, -90000},
