+++ hardware/stepperCalculator/configuration/AxisConfiguration.cs
@@ -2,8 +2,8 @@
 {
     public class AxisConfiguration
     {
-        public float StepsPerMM { get; set; } = 200;
-        public float MaxSpeedPerMM { get; set; } = 20;
+        public decimal StepsPerMM { get; set; } = 200;
+        public decimal MaxSpeedPerMM { get; set; } = 20;
         public int MaxAcceleration { get; set; } = 40;
         public int Jerk { get; set; } = 20;
